Crate bs3

The project structure is divided into four levels

store /Application Layer/

map

vec

value

snapshot /Middle layer with transactional operations/

storage

transaction

model /Cache Layer/

map

Btree<K,Operation>

vec

Btree<usize,Operation>

value

Operation

backend /Storage Layer/

memory

sled

Store Logic

Data is recorded by block height, For example: bytes():Indicates that the data is of type bytes

insert {1:1},{2:2},{3:3}; /*height=0*/
commit; /*height=1,cache is nil*/
{
    bytes({name_space}-ty): bytes(StoreType(3)),
    bytes({name_space}-ch): bytes(StoreHeight(1)),
    bytes({name_space}-kw-{hex(1)}-{:00000000000000000001}): bytes(Operation::Update(1)),
    bytes({name_space}-kw-{hex(2)}-{:00000000000000000001}): bytes(Operation::Update(2)),
    bytes({name_space}-kw-{hex(3)}-{:00000000000000000001}): bytes(Operation::Update(3)),
}
commit; /*height=2*/
{
    bytes({name_space}-ty): bytes(StoreType(3)),
    bytes({name_space}-ch): bytes(StoreHeight(1)),
    bytes({name_space}-kw-{hex(1)}-{:00000000000000000001}): bytes(Operation::Update(1)),
    bytes({name_space}-kw-{hex(2)}-{:00000000000000000001}): bytes(Operation::Update(2)),
    bytes({name_space}-kw-{hex(3)}-{:00000000000000000001}): bytes(Operation::Update(3)),
    bytes({name_space}-ch): bytes(StoreHeight(2)),
}
remove key=1; /*Invalid, cache is nil*/
insert {4:4}; /*height=2*/
commit; /*height=3*/
{
    bytes({name_space}-ty): bytes(StoreType(3)),
    bytes({name_space}-ch): bytes(StoreHeight(1)),
    bytes({name_space}-kw-{hex(1)}-{:00000000000000000001}): bytes(Operation::Update(1)),
    bytes({name_space}-kw-{hex(2)}-{:00000000000000000001}): bytes(Operation::Update(2)),
    bytes({name_space}-kw-{hex(3)}-{:00000000000000000001}): bytes(Operation::Update(3)),
    bytes({name_space}-ch): bytes(StoreHeight(2)),
    bytes({name_space}-ch): bytes(StoreHeight(3)),
    bytes({name_space}-kw-{hex(4)}-{:00000000000000000003}): bytes(Operation::Update(4)),
}

Modules

backend

backend is storage layer

merkle

model

Trait Model constrains the behavior of the cache layer

prelude

Structs

SnapshotableStorage

Snapshotable Storage

Transaction

Enums

Cow

Error

Operation

Traits

DoubleKeyMapStore

Forkable

MapStore

Defining the basic behavior of the map application layer

Store

ValueStore

VecStore

Functions

merkle_key

build merkle root key

Type Definitions

CowBytes

OperationBytes

Result